requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.33.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
cssselect2==0.8.0
etelemetry==0.3.1
eval_type_backport==0.2.2
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.57.0
fpdf==1.7.2
frontend==0.0.3
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.30.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.3.2
narwhals==1.33.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
oauth2client==4.1.3
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
pipeline==0.1.0
plotly==6.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydot==3.0.4
PyDrive==1.3.1
pydyf==0.11.0
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
slugify==0.0.1
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.44.1
streamlit-quill==0.0.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
together==1.5.5
tokenizers==0.21.1
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.51.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
unicode==2.9
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==65.0
webencodings==0.5.1
yarl==1.19.0
zopfli==0.2.3.post1

google-auth-oauthlib~=1.2.1